app/backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via geoip2
aiosignal==1.3.1
    # via aiohttp
alembic==1.14.0
    # via -r requirements.in
attrs==24.2.0
    # via aiohttp
boto3==1.35.58
    # via -r requirements.in
botocore==1.35.58
    # via
    #   boto3
    #   s3transfer
certifi==2024.2.2
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
coverage[toml]==7.5.0
    # via pytest-cov
cryptography==43.0.3
    # via
    #   -r requirements.in
    #   http-ece
    #   py-vapid
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-semantic-conventions
distlib==0.3.8
    # via virtualenv
filelock==3.14.0
    # via virtualenv
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
geoalchemy2==0.15.2
    # via -r requirements.in
geoip2==4.8.0
    # via -r requirements.in
googleapis-common-protos==1.64.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
greenlet==3.0.3
    # via sqlalchemy
grpcio==1.67.1
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
http-ece==1.2.1
    # via -r requirements.in
identify==2.5.36
    # via pre-commit
idna==3.7
    # via
    #   requests
    #   yarl
importlib-metadata==8.0.0
    # via opentelemetry-api
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
luhn==0.2.0
    # via -r requirements.in
mako==1.3.3
    # via alembic
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
maxminddb==2.6.2
    # via geoip2
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
nodeenv==1.9.1
    # via pre-commit
numpy==1.26.4
    # via shapely
opentelemetry-api==1.26.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-instrumentation-threading
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp==1.26.0
    # via -r requirements.in
opentelemetry-exporter-otlp-proto-common==1.26.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.26.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.26.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.47b0
    # via
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-instrumentation-threading
opentelemetry-instrumentation-grpc==0.47b0
    # via -r requirements.in
opentelemetry-instrumentation-sqlalchemy==0.47b0
    # via -r requirements.in
opentelemetry-instrumentation-threading==0.47b0
    # via -r requirements.in
opentelemetry-proto==1.26.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.26.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.47b0
    # via
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-sdk
packaging==24.0
    # via
    #   geoalchemy2
    #   opentelemetry-instrumentation-sqlalchemy
    #   pytest
phonenumbers==8.13.49
    # via -r requirements.in
platformdirs==4.2.2
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.0.1
    # via -r requirements.in
prometheus-client==0.21.0
    # via -r requirements.in
protobuf==4.25.5
    # via
    #   -r requirements.in
    #   googleapis-common-protos
    #   opentelemetry-proto
psycopg2-binary==2.9.10
    # via -r requirements.in
py-vapid==1.9.1
    # via -r requirements.in
pycparser==2.22
    # via cffi
pynacl==1.5.0
    # via -r requirements.in
pytest==8.3.3
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==6.0.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   botocore
pytz==2024.2
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   pre-commit
requests==2.32.3
    # via
    #   -r requirements.in
    #   geoip2
    #   opentelemetry-exporter-otlp-proto-http
    #   stripe
ruff==0.7.3
    # via -r requirements.in
s3transfer==0.10.1
    # via boto3
sentry-sdk==2.18.0
    # via -r requirements.in
shapely==2.0.6
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlalchemy==2.0.36
    # via
    #   -r requirements.in
    #   alembic
    #   geoalchemy2
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.2
    # via -r requirements.in
stripe==11.2.0
    # via -r requirements.in
typing-extensions==4.11.0
    # via
    #   alembic
    #   opentelemetry-sdk
    #   sqlalchemy
    #   stripe
ua-parser==0.18.0
    # via
    #   -r requirements.in
    #   user-agents
urllib3==2.2.1
    # via
    #   botocore
    #   requests
    #   sentry-sdk
user-agents==2.2.0
    # via -r requirements.in
virtualenv==20.26.2
    # via pre-commit
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-instrumentation-threading
yarl==1.11.1
    # via aiohttp
zipp==3.20.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
